lib/vsprintf.c: remove accidental VLA usage
The "sym" calculation is actually a fixed size, but since the max() macro
uses some extensive tricks for safety, it ends up looking like a variable
size.  This replaces max() with a simple max macro which is sufficient for
the calculation of the array size.

Seen with -Wvla. Fixed as part of the directive to remove all VLAs from
the kernel: https://lkml.org/lkml/2018/3/7/621

Link: http://lkml.kernel.org/r/20180307230714.GA20797@beast
Signed-off-by: Kees Cook <keescook@chromium.org>
Cc; "Tobin C. Harding" <me@tobin.cc>
Cc: Jonathan Corbet <corbet@lwn.net>
Cc: Pantelis Antoniou <pantelis.antoniou@konsulko.com>
Cc: Petr Mladek <pmladek@suse.com>
Cc: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>
Cc: "Steven Rostedt (VMware)" <rostedt@goodmis.org>
Cc: "Gustavo A. R. Silva" <gustavo@embeddedor.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>